requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
allure-pytest>=2.13.0
pydantic>=2.5.0
faker>=22.0.0
jsonschema>=4.20.0
python-dotenv>=1.0.0
ruff>=0.1.0
